// src/sdk/types.gen.ts
export type BlankEnum = '';
// ...
export type Incident = {
    readonly uuid: string;
    incident_name?: string | null;
    resident: string;
    readonly resident_uuid: string;
    reported_by?: string | null;
    location?: string | null;
    region?: string | null;
    received_date?: string | null;
    incident_datetime?: string | null;
    agency_name?: string | null;
    pre_incident_notes?: string | null;
    incident_description?: string | null;
    response_action?: string | null;
    send_notification?: boolean;
    status: IncidentStatusEnum;
    readonly lead_uuid: string;
    readonly referral_number: string;
    readonly signature_url: string;
    resident_details: IncidentResidentDetails;
    readonly resident_status: string;
    reported_by_details: IncidentUser;
    readonly notifications: Array<IncidentNotification>;
    readonly comments: Array<IncidentComment>;
    medical_flags?: Array<IncidentMedicalFlag>;
    legal_flags?: Array<IncidentLegalFlag>;
    social_flags?: Array<IncidentSocialFlag>;
    victim_flags?: Array<IncidentVictimFlag>;
    readonly signature: string;
    readonly group_home: string;
    readonly assigned_program_manager: string;
    program_manager_email?: string | null;
    coordinator_email?: string | null;
    readonly reporter_signature_url: string;
    readonly pm_signature_url: string;
    readonly pm_signature: string;
    readonly started_at: string | null;
    readonly submitted_for_review_at: string | null;
    readonly completed_at: string | null;
    pm_review_notes?: string | null;
    pm_program_type?: string | null;
    pm_service_transition?: PmServiceTransitionEnum | BlankEnum | NullEnum | null;
    pm_service_transition_description?: string | null;
    pm_behavior_plan_followed?: PmBehaviorPlanFollowedEnum | BlankEnum | NullEnum | null;
    pm_title?: string | null;
    readonly edits: string;
    created_at?: string;
    readonly updated_at: string;
};

export type IncidentComment = {
    readonly uuid: string;
    comment: string;
    readonly role_name: string;
    readonly created_by_name: string;
    readonly created_at: string;
};

export type IncidentLegalFlag = {
    readonly uuid: string;
    legal_type: string;
};

export type IncidentMedicalFlag = {
    readonly uuid: string;
    medical_type: string;
    medical_other_details?: string;
};

export type IncidentNotification = {
    readonly uuid: string;
    incident: number;
    type: IncidentNotificationTypeEnum;
    readonly user: number | null;
    readonly user_name: string;
    readonly user_email: string;
    notify?: boolean;
    notify_date?: string | null;
    notify_time?: string | null;
    method_of_contact?: string | null;
    by_whom?: string | null;
    created_at?: string;
    readonly updated_at: string;
};

/**
 * * `STAFF` - Staff
 * * `SERVICE_COORDINATOR` - Service Coordinator
 * * `PROGRAM_MANAGER` - Program Manager
 * * `GUARDIAN` - Guardian
 * * `AGENT` - Agent
 * * `NURSING` - Nursing
 * * `ADDITIONAL_SERVICE_PROVIDER` - Additional Service Provider
 */
export type IncidentNotificationTypeEnum = 'STAFF' | 'SERVICE_COORDINATOR' | 'PROGRAM_MANAGER' | 'GUARDIAN' | 'AGENT' | 'NURSING' | 'ADDITIONAL_SERVICE_PROVIDER';

export type IncidentResidentDetails = {
    readonly uuid: string;
    email: string;
    first_name: string;
    last_name?: string | null;
    readonly avatar_url: string;
    readonly lead_uuid: string;
    readonly group_home: string;
    readonly date_of_birth: string;
};

export type IncidentSocialFlag = {
    readonly uuid: string;
    social_type: string;
    social_other_details?: string;
};

/**
 * * `DRAFT` - Draft
 * * `IN_PROGRESS` - In Progress
 * * `PM_REVIEW_PENDING` - Program Manager Review Pending
 * * `COMPLETED` - Completed
 * * `ACKNOWLEDGED` - Acknowledged
 */
export type IncidentStatusEnum = 'DRAFT' | 'IN_PROGRESS' | 'PM_REVIEW_PENDING' | 'COMPLETED' | 'ACKNOWLEDGED';

export type IncidentUser = {
    readonly uuid: string;
    email: string;
    first_name: string;
    last_name?: string | null;
    readonly avatar_url: string;
    readonly role: string;
};

export type IncidentVictimFlag = {
    readonly uuid: string;
    victim_type: string;
};
// ...
export type NullEnum = unknown;
// ...
export type PmBehaviorPlanFollowedEnum = 'YES' | 'NO' | 'N_A';

/**
 * * `YES` - Yes
 * * `NO` - No
 */
export type PmServiceTransitionEnum = 'YES' | 'NO';